pyproject.toml
[project]
name = "guiml"
version = "0.2.0"
authors = [
  { name="Stephan Gocht", email="gocht@mailbox.org" },
]
description = "An xml based GUI framework with property binding and dependency injection."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml",
    "pyglet",
    "pycairo",
    "cairocffi",
    "pangocffi",
    "pangocairocffi",
    "pygobject",
]
license-files = ["LICENSE", "guimlcomponents/base/cairocffi_to_pycairo/LICENSE"]


[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[project.urls]
Homepage = "https://github.com/StephanGocht/guiml"
Issues = "https://github.com/StephanGocht/guiml/issues"
Documentation = "http://guiml.readthedocs.io/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/guiml", "src/guimlcomponents"]
